// src/app/core/api/model/adminUsersListResponse.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminUserRowDto } from './adminUserRowDto';


export interface AdminUsersListResponse { 
    content?: Array<AdminUserRowDto>;
    totalElements?: number;
    totalPages?: number;
}