// src/app/core/api/model/createTutoredUserRequestDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateTutoredUserRequestDto {
  firstName: string;
  lastName: string;
  birthYear?: number;
  phone?: string;
  notesAdmin?: string;
}